Split files and fixed Filter problems
